chore: clear new compiler warnings in klause-ortools and klause-bench

+++ klause-ortools/src/jvmMain/kotlin/com/eignex/klause/ortools/OrToolsSolver.kt
@@ -124,14 +124,14 @@ class OrToolsSolver(override val problem: Problem) : Optimizer<OrToolsParams> {
             val w = obj.boolWeights.getOrElse(b) { 0L }
             if (w != 0L) {
                 args.add(m.boolVars[b])
-                coeffs.add(w.toLong())
+                coeffs.add(w)
             }
         }
         for (i in 0 until problem.numIntVars) {
             val c = obj.intCoefficients.getOrElse(i) { 0L }
             if (c != 0L) {
                 args.add(m.intVars[i])
-                coeffs.add(c.toLong())
+                coeffs.add(c)
             }
         }
         return if (args.isEmpty()) {
